# Conflicts:
#	TheForceEngine/TFE_Jedi/Renderer/RClassic_Float/robj3d_float/robj3dFloat_TransformAndLighting.cpp

+++ TheForceEngine/TFE_Editor/LevelEditor/entity.cpp
@@ -667,8 +667,8 @@ namespace LevelEditor
 		{
 			if (writeEntityDataToString(entity, buffer, 4096))
 			{
-				file.writeBuffer(buffer, strlen(buffer));
-				file.writeBuffer("\r\n", strlen("\r\n"));
+				file.writeBuffer(buffer, (u32)strlen(buffer));
+				file.writeBuffer("\r\n", (u32)strlen("\r\n"));
 			}
 		}
